style(modeling) split cylinder out from cylinderElliptic code
