+++ packages/modeling/src/primitives/index.js
@@ -11,8 +11,8 @@ module.exports = {
   circle: require('./circle'),
   cube: require('./cuboid').cube,
   cuboid: require('./cuboid').cuboid,
-  cylinder: require('./cylinderElliptic').cylinder,
-  cylinderElliptic: require('./cylinderElliptic').cylinderElliptic,
+  cylinder: require('./cylinder'),
+  cylinderElliptic: require('./cylinderElliptic'),
   ellipse: require('./ellipse'),
   ellipsoid: require('./ellipsoid').ellipsoid,
   geodesicSphere: require('./geodesicSphere'),
